secure-ai/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# Document Processing
pymupdf==1.24.2
python-magic-bin==0.4.14

# OCR
onnxruntime==1.18.0
rapidocr-onnxruntime==1.3.22
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0

# NER / ML
transformers==4.40.2
tokenizers==0.19.1
torch==2.2.2

# Security
cryptography==42.0.7

# Database
sqlalchemy==2.0.29

# Observability
prometheus-client==0.20.0
python-json-logger==2.0.7

# Utilities
google-re2==1.1
watchdog==4.0.0
